Fix issues with large 16bit code units
Given a 16bit code units (Fullwidth Left/Right Parenthesis and others)
the optimization in `buildDoubleArrayTrie` doesn't work anymore and we
should revert to the Java implementation's allocation again.

Furthermore, the Java implementation used a `TreeMap<Character, State>`
which is _ordered_ by its key, whereas the port used a
`Dictionary<char, State>` instead which doesn't define any order. This
becomes a problem when iterating over the entries, creating a `List` in
the builder's `fetch` method, then accessing the last item in that list
by index to determine the new Trie size in `insert`.